+++ staticpipes/pipes/collection_records_process.py
@@ -14,14 +14,43 @@ class PipeCollectionRecordsProcess(BasePipe):
 
     Typical uses include with the Jinja2 process, so you can make
     a HTML page for every item in a collection.
+
+    Pass:
+
+    - collection_name
+
+    - processors
+
+    - output_mode. Should be one of "file" (default) or "dir".
+    In "file" mode, the output file for each record is
+    "output_dir/id.output_filename_extension".
+    In "dir" mode, it's "output_dir/id/output_filename"
+
+    - output_dir
+
+    - output_filename_extension
+
+    - output_filename
+
+    - context_key_record_id
+
+    - context_key_record_data
+
+    - context_key_record_class
+
+    - filter_function
+
+
     """
 
     def __init__(
         self,
         collection_name: str,
         processors: list,
+        output_mode: str = "file",
         output_dir=None,
         output_filename_extension="html",
+        output_filename="index.html",
         context_key_record_id: str = "record_id",
         context_key_record_data: str = "record_data",
         context_key_record_class: str = "record",
@@ -29,8 +58,10 @@ class PipeCollectionRecordsProcess(BasePipe):
     ):
         self._collection_name = collection_name
         self._processors = processors
+        self._output_mode = output_mode
         self._output_dir = output_dir or collection_name
         self._output_filename_extension = output_filename_extension
+        self._output_filename = output_filename
         self._context_key_record_id = context_key_record_id
         self._context_key_record_data = context_key_record_data
         self._context_key_record_class = context_key_record_class
@@ -59,9 +90,18 @@ class PipeCollectionRecordsProcess(BasePipe):
             this_context[self._context_key_record_data] = record.get_data()
             this_context[self._context_key_record_class] = record
 
+            if self._output_mode == "file":
+                new_dir = self._output_dir
+                new_filename = record.get_id() + "." + self._output_filename_extension
+            elif self._output_mode == "dir":
+                new_dir = self._output_dir + "/" + record.get_id()
+                new_filename = self._output_filename
+            else:
+                raise Exception("Unrecognised output mode")
+
             process_current_info = ProcessCurrentInfo(
-                self._output_dir,
-                record.get_id() + "." + self._output_filename_extension,
+                new_dir,
+                new_filename,
                 "",
                 prepare=False,
                 build=True,