+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- More robust handling of dir strings
+
 ## [0.3.0] - 2025-06-15
 
 ### Added