+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Build command exits with sys.exit error codes if checks fail, so can use in CI https://github.com/StaticPipes/StaticPipes/issues/54
 - Add check command to CLI. Exits with sys.exit error codes if checks fail, so can use in CI https://github.com/StaticPipes/StaticPipes/issues/54
 - Added `arguments` to data collected in `PipeLoadCollectionPythonDocs` https://github.com/StaticPipes/StaticPipes/issues/26
+- Added `ProcessRemoveHTMLExtension` https://github.com/StaticPipes/StaticPipes/issues/19
 
 ### Fixed